Factor out p_fossil_nuclear computation.
